requirements.txt
Pillow==8.0.1
aiohttp==3.6.2
async-generator==1.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.1.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Cybernator==0.6.2
discord.py==1.7.3
Flask==1.1.2
future==0.18.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.2
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
multidict==4.7.6
pip==20.3.3
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-box==5.2.0
requests==2.7.0
rfc3986==1.4.0
sniffio==1.1.0
soupsieve==2.0.1
tqdm==4.54.0
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.5.1